#8 | Move Manual package to Location
